proxy_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.19.0
python-jose[cryptography]==3.3.0
